+++ src/components/MediaPlayer/MediaPlayer.tsx
@@ -12,7 +12,7 @@ const useStyles = makeStyles(() => ({
   },
   image: {
     width: "100%",
-    height: "100%",
+    height: "99%",
     objectFit: "contain",
   },
   youtube: {