+++ src/features/game/routes/Game.tsx
@@ -48,7 +48,7 @@ const useStyles = makeStyles(() => ({
   },
   image: {
     width: "100%",
-    height: "100%",
+    height: "99%",
     objectFit: "contain",
   },
 }));